requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
aws-dynamodb-parser==0.1.2
boto3==1.18.20
botocore==1.21.20
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.0
future==0.18.2
h11==0.12.0
idna==3.2
jmespath==0.10.0
mangum==0.11.0
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
PyVimeo==1.1.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
starlette==0.14.2
tinydb==4.5.1
tuspy==0.2.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
